fix(lfs): propagate caller cancellation through batch transfers
